backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8

# MongoDB
pymongo==4.10.1
dnspython==2.7.0

# Authentication
PyJWT==2.10.1
bcrypt==4.2.1

# AI / LangChain
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-core==0.3.29
langgraph==0.2.60
google-generativeai==0.8.3
crewai==0.79.4
# Keep crewai-tools optional on Windows unless C++ Build Tools are installed.
# It pulls embedchain/chroma-hnswlib which may require native compilation.

# OCR
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0

# PDF Generation
reportlab==4.2.5
# requests warns if chardet>=6 is installed; reportlab depends on chardet.
chardet<6
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
requests==2.32.3
gunicorn==23.0.0
